Implement argument parsing via argp.h.
This commit implements argument parsing from the command line via
argp.h. So far, I have added only one parameter ('-d', or
'--declaracao') which is used for providing the filename of the
declaration.  I have also marked this argument as optional, so that the
user will still be able to provide the filename without using '-d'.

As an auxiliary support, I enabled the generation of the config.h header
file, so that the program can obtain the package version and the bug
report address directly from configure.ac.