Add comments to configure.ac
I think it is important to make configure.ac nicer to read, because
sometimes one has to hack it and decrypt the code in there.  This commit
achieves this goal by (a) adding the GPL license header, (b) commenting
the important steps of configure.ac, and (c) updating some parameters
from certain macros.