Update/improve the way we check for libraries
Using PKG_CHECK_MODULES is not very neat because the user might want to
be able to provide a different $LDFLAGS, and this would break the build.
Instead, we should be using autoconf's default
AC_CHECK_LIB/AC_SEARCH_LIBS (in our case, I chosed the latter).  This
way, we don't need to modify $LIBS inside configure.ac.

This patch also adds a check for zlib, and updates $CFLAGS accordingly
based on the output of 'pkg-config' for both GNUTLS and zlib libraries.