Fix bug when saving the receipt file using a .DEC filename as template
This commit fixes a bug that happens when the user provides a .DEC
filename to be transmitted, and rnetclient uses this filename to compose
the receipt (.REC) filename to be saved.  I forgot to use "basename" to
extract only the .DEC filename (and strip all the other paths) before using
it to build the .REC full path.  As an addition, this commit also fixes
a small memory leak.  Tested by sending a declaration again.